Omit assets in non-skinnable builds when mod.assets === false
When a module's build config includes the "assets" property set to `false`,
we should skip copying any existing assets folder for that build target.

However, if that module's meta.json includes the "skinnable" property set
to true, the assets directory will still be copied, as it is required by
the skinning pattern.

Fixes #87