move `args.unshift(util.shifter)` out of walk.js's run() loop
The path 'shifter/bin/shifter' was prepended to the spawn argument list
on each iteration of run(), which can be called many times in walk.js.
This grew the argument list. The extra args are ultimately ignored, but
this small fix prevents non-parsed arg pollution (i.e. the values nopt
returns in the parsed.remain property).